worked on custom commander

+++ The Boring Commander/boring.h
@@ -8,5 +8,5 @@
 #define BORING_H
 	void execute(char** command); //executes the given command, calls show() to display statistics
 
-	void show(long elaspedTime, int pageFaults, int reclaimedPageFaults);
+	void show(long elaspedTime, long pageFaults, long reclaimedPageFaults);
 #endif
